+++ src/main/kotlin/io/unthrottled/doki/actions/MotivatorPromotion.kt
@@ -6,6 +6,6 @@ import io.unthrottled.doki.promotions.MotivatorPluginPromotion
 
 class MotivatorPromotion : AnAction() {
   override fun actionPerformed(e: AnActionEvent) {
-    MotivatorPluginPromotion.runPromotion { }
+    MotivatorPluginPromotion.runPromotion({}) { }
   }
 }